+++ src/main/java/com/omteam/omt/report/repository/DailyAnalysisRepository.java
@@ -1,6 +1,6 @@
-package com.omteam.omt.character.repository;
+package com.omteam.omt.report.repository;
 
-import com.omteam.omt.character.domain.DailyAnalysis;
+import com.omteam.omt.report.domain.DailyAnalysis;
 import java.time.LocalDate;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
